requirements.txt
fastapi
python-dotenv
uvicorn
colorama
pysqlite3
bcrypt